refactoring project layout refactoring startup component to more generic wait component
