+++ src/components/index.ts
@@ -1,5 +1,3 @@
-export * from './state';
-export * from './actions';
-export * from './context';
-export * from './AppComponent';
-export * from './startup/Startup';
+export * from './wait/Wait';
+export * from './wait/Startup';
+export * from './demo/Demo';
