+++ src/epics/context.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import { ActionsObservable } from 'redux-observable';
 import { DangerZone } from 'expo';
-import { RequestAction, actionTypes, getRequestAction, LocalePayload } from '../components';
+import { RequestAction, actionTypes, getRequestAction, LocalePayload } from '../framework';
 
 export function loadLocale(actions: ActionsObservable<RequestAction>): Observable<RequestAction<LocalePayload>> {
   return actions
