+++ test/app.test.tsx
@@ -13,16 +13,4 @@ describe('App', () => {
     // Assert
     expect(result).toMatchSnapshot();
   });
-
-  // it('binds to the redux store', () => {
-  //   // Arrange
-  //   const component = (<App />);
-
-  //   // Act
-  //   const result = shallow(component);
-  //   const store = result.prop('store');
-
-  //   // Assert
-  //   expect(store).toBeTruthy();
-  // });
 });
