+++ test/components/demo/Demo.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { Button } from 'react-native';
 import { createStoreProps } from '../../storeUtils';
-import { Store, actionTypes } from '../../../src/components';
+import { Store, actionTypes } from '../../../src/framework';
 import { Demo } from '../../../src/components/demo/Demo';
 
 const defaultState: Partial<Store> = {
