+++ test/screens/Login.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { createStoreProps } from '../storeUtils';
 import { Login } from '../../src/screens/Login';
-import { Store } from '../../src/components';
+import { Store } from '../../src/framework';
 
 const defaultState: Partial<Store> = {
   login: {},
